+++ core/src/main/java/nl/nn/adapterframework/http/rest/ApiListener.java
@@ -60,7 +60,7 @@ public class ApiListener extends PushingListenerAdapter implements HasPhysicalDe
 
 	private final @Getter(onMethod = @__(@Override)) String domain = "Http";
 	private @Getter String uriPattern;
-	private @Getter boolean updateEtag = AppConstants.getInstance().getBoolean("api.etag.enabled", true);
+	private @Getter boolean updateEtag = AppConstants.getInstance().getBoolean("api.etag.enabled", false);
 	private @Getter String operationId;
 
 	private @Getter HttpMethod method = HttpMethod.GET;