+++ RELEASES.md
@@ -32,6 +32,7 @@ Upcoming (7.9)
     `servlet.WebServices11.enabled=true` or 
     `servlet.BrowserBinding.enabled=true`
 - Some API endpoints have been deprecated. Users are encouraged to change over to the new API, however in order to restore the deprecated functionality the property 'iaf-api.allowDeprecated' can be set to true.
+- ApiListener eTag generation has been disabled by default, set api.etag.enabled=true to enable default etag generation.
 
 
 7.8-RC1